merge main into teamProfilePage_Muneeb

See merge request team-projects-2023-24/team14!13

+++ src/main/java/team/bham/web/rest/UserProfileResource.java
@@ -15,9 +15,14 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
+import team.bham.domain.User;
 import team.bham.domain.UserProfile;
 import team.bham.repository.UserProfileRepository;
+import team.bham.repository.UserRepository;
+import team.bham.security.SecurityUtils;
+import team.bham.service.UserProfileService;
 import team.bham.web.rest.errors.BadRequestAlertException;
+import team.bham.web.rest.errors.EmailAlreadyUsedException;
 import tech.jhipster.web.util.HeaderUtil;
 import tech.jhipster.web.util.ResponseUtil;
 
@@ -31,15 +36,31 @@ public class UserProfileResource {
 
     private final Logger log = LoggerFactory.getLogger(UserProfileResource.class);
 
+    private static class UserProfileResourceException extends RuntimeException {
+
+        private UserProfileResourceException(String message) {
+            super(message);
+        }
+    }
+
     private static final String ENTITY_NAME = "userProfile";
 
     @Value("${jhipster.clientApp.name}")
     private String applicationName;
 
-    private final UserProfileRepository userProfileRepository;
+    private final UserProfileService userProfileService;
 
-    public UserProfileResource(UserProfileRepository userProfileRepository) {
+    private final UserProfileRepository userProfileRepository;
+    private final UserRepository userRepository;
+
+    public UserProfileResource(
+        UserProfileService userProfileService,
+        UserProfileRepository userProfileRepository,
+        UserRepository userRepository
+    ) {
+        this.userProfileService = userProfileService;
         this.userProfileRepository = userProfileRepository;
+        this.userRepository = userRepository;
     }
 
     /**
@@ -55,6 +76,11 @@ public class UserProfileResource {
         if (userProfile.getId() != null) {
             throw new BadRequestAlertException("A new userProfile cannot already have an ID", ENTITY_NAME, "idexists");
         }
+        long userId = userProfileService.getUserId();
+        if (userProfileRepository.existsById(userId)) {
+            throw new UserProfileResource.UserProfileResourceException("Profile Already exists");
+        }
+        userProfile.setId(userId);
         UserProfile result = userProfileRepository.save(userProfile);
         return ResponseEntity
             .created(new URI("/api/user-profiles/" + result.getId()))
@@ -89,6 +115,11 @@ public class UserProfileResource {
             throw new BadRequestAlertException("Entity not found", ENTITY_NAME, "idnotfound");
         }
 
+        long userId = userProfileService.getUserId();
+        if (!Objects.equals(userId, userProfile.getId())) {
+            throw new UserProfileResource.UserProfileResourceException("Not Authorised"); //replace with 403 response
+        }
+
         UserProfile result = userProfileRepository.save(userProfile);
         return ResponseEntity
             .ok()
@@ -113,6 +144,7 @@ public class UserProfileResource {
         @NotNull @RequestBody UserProfile userProfile
     ) throws URISyntaxException {
         log.debug("REST request to partial update UserProfile partially : {}, {}", id, userProfile);
+
         if (userProfile.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
@@ -122,6 +154,10 @@ public class UserProfileResource {
 
         if (!userProfileRepository.existsById(id)) {
             throw new BadRequestAlertException("Entity not found", ENTITY_NAME, "idnotfound");
+        }
+        long userId = userProfileService.getUserId();
+        if (!Objects.equals(userId, userProfile.getId())) {
+            throw new UserProfileResource.UserProfileResourceException("Not Authorised"); //replace with 403 response
         }
 
         Optional<UserProfile> result = userProfileRepository
@@ -202,6 +238,10 @@ public class UserProfileResource {
      */
     @DeleteMapping("/user-profiles/{id}")
     public ResponseEntity<Void> deleteUserProfile(@PathVariable Long id) {
+        long userId = userProfileService.getUserId();
+        if (!Objects.equals(userId, id)) {
+            throw new UserProfileResource.UserProfileResourceException("Not Authorised"); //replace with 403 response
+        }
         log.debug("REST request to delete UserProfile : {}", id);
         userProfileRepository.deleteById(id);
         return ResponseEntity
